pyproject.toml
[project]
name = "chicagoparks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "flask>=3.1.0",
    "flask-session>=0.8.0",
    "httpx>=0.28.1",
    "jellyfish>=1.2.0",
    "lxml>=5.3.2",
    "mysql>=0.0.3",
    "mysql-connector-python>=9.3.0",
    "mysqlclient>=2.2.7",
    "pandas>=2.2.3",
    "playwright>=1.51.0",
    "python-dotenv>=1.1.0",
    "shapely>=2.1.0",
    "sodapy>=2.2.0",
]
